requirements.txt
nba_api>=1.4.1
pandas>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
gspread>=5.12.0
google-auth>=2.23.0
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
